fix: handle native menu actions on macOS
